feat: Created the components TodoForm and TodoList. refactor: Updated the App.jsx to use the new components
