+++ src/components/TodoForm.jsx
@@ -1,0 +1,10 @@
+function TodoForm() {
+  return (
+    <form action="">
+      <label htmlFor="todoTile"> Todo</label>
+      <input id="todoTItle" type="text" />
+      <button> Add Todo</button>
+    </form>
+  );
+}
+export default TodoForm;
